Pull request #1094: fix overflow in memory computation
Merge in CLOUD/terraform-provider-yandex-mirror from fix-memory-size-overflow to master

Squashed commit of the following:

commit 8d7fe04fa7e0d61b9dcaaeaf9be86d630cc54fa0
Author: Sergey Dubovtsev <dubovtsev@yandex-team.ru>
Date:   Tue Jan 23 13:03:54 2024 +0300

    fix overflow in memory computation

+++ yandex/resource_yandex_function.go
@@ -521,7 +521,7 @@ func expandLastVersion(d *schema.ResourceData) (*functions.CreateFunctionVersion
 	versionReq.Runtime = d.Get("runtime").(string)
 	versionReq.Entrypoint = d.Get("entrypoint").(string)
 
-	versionReq.Resources = &functions.Resources{Memory: int64(int(datasize.MB.Bytes()) * d.Get("memory").(int))}
+	versionReq.Resources = &functions.Resources{Memory: int64(datasize.MB.Bytes()) * int64(d.Get("memory").(int))}
 	if v, ok := d.GetOk("execution_timeout"); ok {
 		i, err := strconv.ParseInt(v.(string), 10, 64)
 		if err != nil {
